Move YaccGrammar (and associated bits) into their own module.
yacc/mod.rs was starting to get rather full.
